--- Sources/muterCore/CLICommands/RunCommand/Steps/ExcludedMutationPointsDetector.swift
@@ -1,37 +1,0 @@
-import Foundation
-import SwiftSyntax
-
-// Currently supports only line comments (in block comments, would need to detect in which actual line the skip marker appears - and if it isn't the first or last line, it won't contain code anyway)
-final class ExcludedMutationPointsDetector: SyntaxAnyVisitor, PositionDiscoveringVisitor {
-    var positionsOfToken: [MutationPosition] = []
-    
-    private let muterSkipMarker = "muter:skip"
-    
-    private let sourceFileInfo: SourceFileInfo
-    
-    init(configuration: MuterConfiguration?, sourceFileInfo: SourceFileInfo) {
-        self.sourceFileInfo = sourceFileInfo
-    }
-    
-    override func visitAnyPost(_ node: Syntax) {
-        node.leadingTrivia.map { leadingTrivia in
-            if leadingTrivia.containsLineComment(muterSkipMarker) {
-                positionsOfToken.append(
-                    node.mutationPosition(with: sourceFileInfo)
-                )
-            }
-        }
-    }
-}
-
-private extension SwiftSyntax.Trivia {
-    func containsLineComment(_ comment: String) -> Bool {
-        return contains { piece in
-            if case .lineComment(let commentText) = piece {
-                return commentText.contains(comment)
-            } else {
-                return false
-            }
-        }
-    }
-}
